+++ nifti2/nifti_tool.h
@@ -305,7 +305,7 @@ NI2_API int disp_nifti1_extension( const char *mesg, nifti1_extension * ext, int
 NI2_API int disp_field       (const char *mesg,field_s *fieldp,void *str,int nfields,int header);
 NI2_API int disp_field_s_list(const char * mesg, field_s *, int nfields);
 NI2_API int disp_nt_opts     ( const char *mesg, nt_opts * opts);
-NI2_API int disp_raw_data    (void * data, int type, int nvals, char space,int newline);
+NI2_API int disp_raw_data    (void * data, int type, int64_t nvals, char space,int newline);
 NI2_API int fill_cmd_string  (nt_opts * opts, int argc, char * argv[]);
 NI2_API int fill_field       (field_s *fp, int type, int offset, int num, const char *name);
 NI2_API int fill_hdr1_field_array(field_s * nh_fields);
@@ -327,7 +327,7 @@ NI2_API int write_hdr2_to_file(nifti_2_header * nhdr, const char * fname);
 /* wrappers for nifti reading functions (allow MAKE_IM) */
 NI2_API nifti_image    * nt_image_read (nt_opts * opts, const char * fname,
                                         int read_data, int make_ver);
-NI2_API nifti_image    * nt_read_bricks(nt_opts * opts, char * fname, int len,
+NI2_API nifti_image    * nt_read_bricks(nt_opts * opts, char * fname, int64_t len,
                                         int64_t * list, nifti_brick_list * NBL);
 NI2_API void * nt_read_header(const char * fname, int * nver, int * swapped, int check,
                               int new_datatype, int64_t new_dim[8]);
